Add files via upload
Base classes that implement a memory pool (object pool) design pattern. It can be used by applications that allocates huge number of objects from the same class, in which each object has a maximum known size.
CMemoryPoolItem is the base class for any object that needs to be managed in a memory pool. Any class can inherit this base class to allow its objects to be managed by the memory pool manager.
CFullPathMemoryPoolManager is the base class of the memory pool manager. This class must be inherited and the child class must implement its AllocateMemoryPoolItem method to instantiate an object from the class that inherits from CMemoryPoolItem.